feat: xp gain from literal

+++ src/client/kotlin/tech/thatgravyboat/skyblockapi/api/events/info/ActionBarWidgetChangeEvent.kt
@@ -45,7 +45,7 @@ class DrillActionBarWidgetChangeEvent(val current: Int, val max: Int, old: Strin
 class SkillXpPercentActionBarWidgetChangeEvent(val amount: Float, val skill: Skill?, val percent: Float, old: String, new: String) :
     ActionBarWidgetChangeEvent(ActionBarWidget.SKILL_XP_PERCENT, old, new)
 
-class SkillXpLiteralActionBarWidgetChangeEvent(val amount: Float, val skill: Skill?, val current: Int, val needed: Int, old: String, new: String) :
+class SkillXpLiteralActionBarWidgetChangeEvent(val amount: Float, val skill: Skill?, val current: Long, val needed: Long, old: String, new: String) :
     ActionBarWidgetChangeEvent(ActionBarWidget.SKILL_XP_LITERAL, old, new)
 
 enum class ActionBarWidget {
